# Add top-level compile time to at-time report.