@johnnychen94 2020-10-08
Would it be okay to just add those two metrics into one? I guess normal users would just get confused about what "top-level compilation time" is and whether it is noteworthy. Adding some explanation of it in the docstring would be helpful, too.