@vchuravy 2020-10-08
> Would it be okay to just add those two metrics into one? 

We thought about this. But the goal of users with `@time` is to investigate runtime, reporting the percentage of that spend on compile-time is useful. The value add that this brings is that it can happen that the top-level compilation time can easily exceed and be perceptible by the user the actual runtime. So you would notice `@time` taking significant time and the reported time being 0. 

That behavior is confusing to novices and experts alike, and by adding a metric that tracks this we hope to at-least give the option for people to realize what is happening.