@yuyichao 2020-10-08
If you want to include all evaluation time, I think this should just be a REPL feature instead. Without major change at a lot of different levels, it's basically impossible to do this correctly. Implementing it in the REPL should be the easiest and most reliable way to get the full timing without any effect on the use in other cases.