@NHDaly 2020-10-08
> If you want to include all evaluation time, I think this should just be a REPL feature instead. Without major change at a lot of different levels, it's basically impossible to do this correctly. Implementing it in the REPL should be the easiest and most reliable way to get the full timing without any effect on the use in other cases.


Thanks, that makes sense.

My main concern is that the metric added in https://github.com/JuliaLang/julia/pull/37678#discussion_r495507457 is somewhat confusing, because we report a percentage of the runtime as compilation time, but indeed there may have been significantly more compilation time than you observe.

I think that the decision you all came to there to report only a percentage, and not an absolute value, for this exact reason, is the right decision. But it would be ideal if we could also include the top-level compilation time so that this doesn't further cement people's belief that all the compilation time is indeed being measured.

I'm not sure if it's possible without major changes, but that would be my hope! <3