@chriselrod 2020-11-28
I'm fairly ignorant of the internal workings of the compiler or REPL, but from that position it does seem like the REPL would be a reliable place to implement such a feature.
I occasionally find myself copy and pasting:
```julia
julia> t = time()

julia> @time # code to run

julia> time() - t
```
to get an actual idea of the total time it took for `# code to run`. I imagine the REPL would be the easiest place to do something like this automatically, without risking having any compilation take place before the first call to `time()` (or `time_ns()`).