@IanButterworth 2021-02-20
@NHDaly I only just saw your followup commits on this. Happy to hear there might be a fix! :) 
Do you still think it's viable?