@NHDaly 2022-01-02
Hey ian, sorry i didn't reply for a long time. I don't exactly remember the details now, but it looks like i _did_ think it was viable... 🤔 It would probably be worth checking to see if this works and if there are any downsides, because i think it would be a nice improvement, yeah!