@IanButterworth 2022-01-02
Tbh after sitting with the new comp % feature for some time, I think the current report may be fine. The main benefit of adding it is to highlight that "the majority of the execution time was compilation", and for that I don't believe it needs to be more precise or complete than it currently is (especially after the much needed improvements to avoid illogical %'s).

For that reason in the original PR the decision was made to steer away from providing numeric measures of compilation time, and just keep it as a % print. 

I also think the REPL has a layer of top-level overhead that's ok to hide, generally?

I guess I just feel like the current implementation meets the problem spec? If not, perhaps a new problem spec could be defined?

I'd be happy to discuss the pros and cons in a call or something.