@NHDaly 2022-01-02
👍 makes sense 👍 yeah, after it's soaked in, i'm happier with the original, too. Thanks!